Add in memory db

+++ Database/ApplicationDbExtension.cs
@@ -9,21 +9,26 @@ public static class ApplicationDbExtension
     public static void RegisterDatabase(this IServiceCollection services, IConfiguration config)
     {
         var connectionString = config.GetConnectionString("DefaultConnection");
-
+        
         if (string.IsNullOrWhiteSpace(connectionString))
-            throw new InvalidOperationException("Connection string not found.");
-
-        services.AddDbContext<ApplicationDbContext>(options =>
-            options.UseNpgsql(connectionString));
+            services.AddDbContext<ApplicationDbContext>(options =>
+                options.UseInMemoryDatabase("LicenseManager"));
+        else
+            services.AddDbContext<ApplicationDbContext>(options =>
+                options.UseNpgsql(connectionString));
     }
 
     public static void MigrateDatabase(this IServiceProvider services)
     {
         using var scope = services.CreateScope();
         var db = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-        db.Database.Migrate();
+
+        if (db.Database.IsRelational())
+            db.Database.Migrate();
+        else
+            db.Database.EnsureCreated();
     }
-    
+
     public static void SeedDatabase(this IServiceProvider services)
     {
         using var scope = services.CreateScope();
@@ -32,13 +37,14 @@ public static class ApplicationDbExtension
 
         if (db.Users.Any()) return;
 
-        db.Database.ExecuteSqlRaw(@"
-            TRUNCATE TABLE ""Seats"", 
-                           ""Licenses"", 
-                           ""LicenseGroups"", 
-                           ""Users""
-            RESTART IDENTITY CASCADE;
-        ");
+        if (db.Database.IsRelational())
+            db.Database.ExecuteSqlRaw(@"
+                TRUNCATE TABLE ""Seats"",
+                               ""Licenses"",
+                               ""LicenseGroups"",
+                               ""Users""
+                RESTART IDENTITY CASCADE;
+            ");
         
         new UserSeeder().Seed(db);
         new GroupSeeder().Seed(db);
